3.4.10. Adpated to pvr addon api 5.2.1 (3.4.9 was just to make it compile again).

+++ src/tvheadend/Subscription.h
@@ -93,7 +93,7 @@ namespace tvheadend
      * @param time timestamp to seek to
      * @return false if the command failed, true otherwise
      */
-    bool SendSeek(int time);
+    bool SendSeek(double time);
 
     /**
      * Change the subscription speed on the backend
